backend/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.1
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.10.0
